refactor(web): extract StorageForm component for reuse by the recover wizard
